Fibonacci with Trampoline

+++ src/main/java/code/kata/fibo/FiboEnd.java
@@ -4,16 +4,16 @@ import static java.lang.String.format;
 
 public class FiboEnd {
 
-    public int fibonacci(int index) {
+    public Long fibonacci(int index) {
         long startTime = System.nanoTime();
-        int fibo = compute(index, 1, 1);
+        Long fibo = compute(index, 1L, 1L);
         long endTime = System.nanoTime();
         System.out.println(format("Computation of fibonacci(%d) took %d ns", index, (endTime - startTime)));
 
         return fibo;
     }
 
-    private int compute(int index, int fibo_nMinus2, int fibo_nMinus1) {
+    private Long compute(int index, Long fibo_nMinus2, Long fibo_nMinus1) {
 
         if(index == 0 || index == 1) {
             return fibo_nMinus1;
